Refactor category component tests: streamline using directives, update namespaces for consistency, and improve readability in CreateTests, DetailsTests, EditTests, and ListTests.
